Move basic components to components
