+++ src/components/LimitNotice.tsx
