+++ src/components/ProgressPanel.tsx
